backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.0.0
django-filter==24.3